requirements-ingest.txt
google-generativeai>=0.8.3
pinecone-client>=5.0.0
pdfplumber>=0.11.4
tqdm>=4.66.1
